refactor: move weight-based tools to new tac_weighting module
